+++ src/components/assign1/ImageBox.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-function ImageBox(props) {
+function ImageBox() {
   return (
     <Container>
       <Image />
