+++ src/pages/Result.jsx
@@ -1,13 +1,19 @@
 import React from 'react';
 import styled from 'styled-components';
+import Nav from '../components/assign1/Nav';
 import ImageBox from '../components/assign1/ImageBox';
 
 function Result(props) {
   return (
-    <div>
+    <Container>
+      <Nav />
       <ImageBox />
-    </div>
+    </Container>
   );
 }
-
+const Container = styled.div`
+  width: 100%;
+  height: 100%;
+  background-color: #fff;
+`;
 export default Result;
